+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/main/java/com/liferay/portal/configuration/extender/internal/FactoryConfigurationDescription.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.configuration.extender;
+package com.liferay.portal.configuration.extender.internal;
 
 import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescription;
 import com.liferay.portal.kernel.util.StringBundler;
