+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/testIntegration/java/com/liferay/portal/configuration/extender/internal/test/ConfiguratorExtensionTest.java
@@ -15,9 +15,9 @@
 package com.liferay.portal.configuration.extender.internal.test;
 
 import com.liferay.arquillian.deploymentscenario.annotations.BndFile;
-import com.liferay.portal.configuration.extender.FactoryConfigurationDescription;
-import com.liferay.portal.configuration.extender.SingleConfigurationDescription;
 import com.liferay.portal.configuration.extender.internal.ConfiguratorExtension;
+import com.liferay.portal.configuration.extender.internal.FactoryConfigurationDescription;
+import com.liferay.portal.configuration.extender.internal.SingleConfigurationDescription;
 import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescription;
 import com.liferay.portal.configuration.extender.internal.api.ConfigurationDescriptionFactory;
 import com.liferay.portal.configuration.extender.internal.api.NamedConfigurationContent;
